sched/fair: turn off EAS_USE_NEED_IDLE sched feature
This feature disables the standard EAS task placement behavior for
prefer_idle tasks, causing them to prefer smaller idle CPUs instead of
the first idle CPU they find. For now let's use the  well-understood
EAS behavior we've used on previous devices, and revisit turning this
feature on if we decide it's an effective way to save power.

Bug: 78782859
Test: sched stats show prefer_idle path being taken
Change-Id: I1b7ff4865f54aa978d900cad5c3bfd69f1b71ce4
Signed-off-by: Connor O'Brien <connoro@google.com>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>